refactor: field product
