+++ src/components/product/index.tsx
@@ -5,18 +5,7 @@ import {Image} from "react-bootstrap";
 import ShoppingCartIcon from "@mui/icons-material/ShoppingCart";
 import LabelDiscount from "./LabelDiscount";
 import LabelNew from "./LabelNew";
-import {ObjectId} from "mongodb";
-
-export interface ProductProps {
-    _id: ObjectId,
-    sale?: boolean,
-    new?: boolean,
-    discount?: number,
-    imagePath: string,
-    name: string,
-    price: number,
-    category: number
-}
+import ProductProps from "../../type/product.type";
 
 export default function Product(props: ProductProps) {
     // Create our number formatter.
@@ -38,7 +27,7 @@ export default function Product(props: ProductProps) {
                 </Stack>
                 <Box className={'object-fit-cover position-absolute top-0 start-0 z-0'}
                      sx={{width: '306px', height: '100%'}}>
-                    <Image className={`${styles.image_hover}`} src={props.imagePath} alt={props.name}
+                    <Image className={`${styles.image_hover}`} src={props.model[0].pathImageColor} alt={props.name}
                            style={{width: '100%'}}/>
                 </Box>
             </Box>
